requirements.txt
asgiref==3.6.0
asttokens==2.1.0
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.2
django-heroku==0.3.1
djangorestframework==3.14.0
entrypoints==0.4
executing==1.2.0
fonttools==4.37.1
future==0.18.2
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
gwebsocket==0.9.9
httpie==3.2.1
idna==3.3
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
jupyter_client==7.4.7
jupyter_core==5.0.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mercadopago==2.2.0
mpi4py==3.1.4
multidict==6.0.2
nest-asyncio==1.5.6
numpy==1.23.2
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.4
prompt-toolkit==3.0.32
psutil==5.9.4
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2==2.9.3
pure-eval==0.2.2
pycodestyle==2.9.0
pycparser==2.21
pygame==2.1.2
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
pyswarm==0.6
pyswarms==1.3.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
scipy==1.9.3
six==1.16.0
spa==0.10.1
sqlparse==0.4.2
stack-data==0.6.1
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.5.0
tzdata==2022.1
urllib3==1.26.12
utc==0.0.3
wcwidth==0.2.5
Werkzeug==2.2.2
whitenoise==6.2.0
zope.event==4.5.0
zope.interface==5.5.1
